+++ Tasks/src/graphs/DirectedEdge.kt
@@ -1,4 +1,4 @@
 package graphs
 
-class DirectedEdge(val from: DiGraphNode?, val to: DiGraphNode?, val weight: Int) {
+class DirectedEdge(val from: DiGraphNode?, val to: DiGraphNode?, val weight: Double) {
 }
